### 2025-05-17 #8209 @editorialbot
[PRE REVIEW]: hashin_shtrikman_mp: a package for the optimal design and discovery of multi-phase composite materials
**Submitting author:** @carlacupcake (<a href="http://orcid.org/0009-0004-0646-6386">Carla Becker</a>)
**Repository:** https://github.com/carlacupcake/hashin_shtrikman_mp.git
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
