### 2025-05-17 #8209 @Kevin-Mattheus-Moerman
@carlacupcake, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Try to address the above paper compilation error. You can call `@editorialbot generate pdf` to try compilation here again.
