### 2025-05-26 c926bbc
Added DOIs to Refs.bib and paper.bib
