+++ paper/latex/Refs.bib
@@ -1,96 +1,99 @@
 @article{zohdi2016macromicro,
-  title={An explicit macro-micro phase-averaged stress correlation for particle-enhanced composite materials in loaded structures},
-  author={Zohdi, TI},
-  journal={International Journal of Engineering Science},
-  volume={109},
-  pages={1--13},
-  year={2016},
-  publisher={El Sevier}
+  author    = {Zohdi, TI},
+  doi       = {10.1016/j.ijengsci.2016.09.005},
+  journal   = {International Journal of Engineering Science},
+  pages     = {1--13},
+  publisher = {El Sevier},
+  title     = {An explicit macro-micro phase-averaged stress correlation for particle-enhanced composite materials in loaded structures},
+  volume    = {109},
+  year      = {2016},
 }
 
 @book{zohdi2012electromagnetic,
-  title={Electromagnetic properties of multiphase dielectrics: a primer on modeling, theory and computation},
-  author={Zohdi, Tarek I},
-  volume={64},
-  year={2012},
-  publisher={Springer Science \& Business Media}
+  author    = {Zohdi, Tarek I},
+  doi       = {10.1007/978-3-642-28427-4},
+  publisher = {Springer Science \& Business Media},
+  title     = {Electromagnetic properties of multiphase dielectrics: a primer on modeling, theory and computation},
+  volume    = {64},
+  year      = {2012},
 }
 
 @article{hashin1962variational,
-  title={A variational approach to the theory of the effective magnetic permeability of multiphase materials},
-  author={Hashin, Zvi and Shtrikman, Shmuel},
-  journal={Journal of applied Physics},
-  volume={33},
-  number={10},
-  pages={3125--3131},
-  year={1962},
-  publisher={American Institute of Physics}
+  author    = {Hashin, Zvi and Shtrikman, Shmuel},
+  doi       = {10.1063/1.1728579},
+  journal   = {Journal of applied Physics},
+  number    = {10},
+  pages     = {3125--3131},
+  publisher = {American Institute of Physics},
+  title     = {A variational approach to the theory of the effective magnetic permeability of multiphase materials},
+  volume    = {33},
+  year      = {1962},
 }
 
-@article{Myhill2023, 
-    doi = {10.21105/joss.05389}, 
-    url = {https://doi.org/10.21105/joss.05389}, 
-    year = {2023}, 
-    publisher = {The Open Journal}, 
-    volume = {8}, 
-    number = {87}, 
-    pages = {5389}, 
-    author = {Robert Myhill and Sanne Cottaar and Timo Heister and Ian Rose and Cayman Unterborn and Juliane Dannberg and Rene Gassmoeller}, 
-    title = {BurnMan -- a Python toolkit for planetary geophysics, geochemistry and thermodynamics}, journal = {Journal of Open Source Software}
+@article{Myhill2023,
+    author    = {Robert Myhill and Sanne Cottaar and Timo Heister and Ian Rose and Cayman Unterborn and Juliane Dannberg and Rene Gassmoeller},
+    doi       = {10.21105/joss.05389},
+    number    = {87},
+    pages     = {5389},
+    publisher = {The Open Journal},
+    title     = {BurnMan -- a Python toolkit for planetary geophysics, geochemistry and thermodynamics}, journal = {Journal of Open Source Software},
+    url       = {https://doi.org/10.21105/joss.05389},
+    volume    = {8},
+    year      = {2023},
 }
 
 @article{ZARE2017176,
[…]
