### 2025-05-26 16def78
Reformatted figure and equation labels in markdown
