+++ paper/markdown/paper.md
@@ -35,12 +35,12 @@ affiliations:
  - name: Energy Technologies Area, Lawrence Berkeley National Laboratory, Berkeley, CA 94720, USA
    index: 3
  - name: Materials Sciences Division, Lawrence Berkeley National Laboratory, Berkeley, California, United States of America
-   index: 3
+   index: 4
 date: 31 March 2025
 bibliography: paper.bib
 
 ---
-# [Summary]{label="sec:summary"}
+# Summary
 
 `hashin_shtrikman_mp` is a tool for composites designers who have
 desired composite properties in mind, but who do not yet have an
@@ -58,7 +58,7 @@ machine-learning and iv) agnosticism to underlying microstructure, and
 ultimate engineering application, make it a tool with much broader
 applications than its predecessors.
 
-# [Statement of Need]{label="sec:need"}
+# Statement of need
 
 Composites are ubiquitous in engineering due to their tunability and
 enhanced material properties as compared to their individual
@@ -100,9 +100,9 @@ candidate theoretical materials, then searches for real materials in the
 Materials Project database with properties close to the recommended
 constituents.
 
-# [Underlying Theory]{label="sec:theory"}
+# Underlying theory
 
-## [Estimate effective composite properties with the Hashin-Shtrikman bounds]{label="subsec:HS-bounds"}
+## Estimate effective composite properties with the Hashin-Shtrikman bounds
 
 When designing composites, simple volume-weighted linear combinations of
 constituent material properties do not yield accurate approximations of
@@ -123,27 +123,30 @@ ordered from least to greatest where
 $y_{1} \leq y_{2} \leq \dotsb \leq y_{n}$ with corresponding volume
 fractions sum to unity $v_{1} + v_{2} + \dotsb + v_{n} = 1$:
 
-$$y_{1} + \frac{A_{1}}{1 - \alpha_{1}A_{1}}
+\begin{equation}\label{eqn:gen_ineq}
+y_{1} + \frac{A_{1}}{1 - \alpha_{1}A_{1}}
 = y^{*,-} \leq y^{*} \leq y^{*,+} = 
 y_{n} + \frac{A_{n}}{1 - \alpha_{n}A_{n}}
-\label{eqn:gen_ineq}$$
+\end{equation}
 
-where $$\alpha_{1} = \frac{1}{3y_{1}}
+where
+\begin{equation}\label{eqn:gen_alphas}
+\alpha_{1} = \frac{1}{3y_{1}}
 \quad \text{and} \quad
 \alpha_{n} = \frac{1}{3y_{n}},
-\label{eqn:gen_alphas}$$
+\end{equation}
 
 and
-$$A_{1} = \sum\limits_{i=2}^{n} \frac{v_{i}}{\frac{1}{y_{i} - y_{1}} + \alpha_{1}}
+\begin{equation}\label{eqn:gen_As}
+A_{1} = \sum\limits_{i=2}^{n} \frac{v_{i}}{\frac{1}{y_{i} - y_{1}} + \alpha_{1}}
 \quad \text{and} \quad
 A_{n} = \sum\limits_{i=1}^{n-1} \frac{v_{i}}{\frac{1}{y_{i} - y_{n}} + \alpha_{n}}.
-\label{eqn:gen_As}$$
+\end{equation}
 
 ![An example of a quasi-isotropic, quasi-homogeneous 3-phase
[…]
