### 2025-05-26 6b9c1af
Paper compiles - fixing small typos
