+++ paper/latex/Refs.bib
@@ -1,5 +1,5 @@
 @article{zohdi2016macromicro,
-  author    = {Zohdi, TI},
+  author    = {Zohdi, T. I.},
   doi       = {10.1016/j.ijengsci.2016.09.005},
   journal   = {International Journal of Engineering Science},
   pages     = {1--13},
@@ -10,7 +10,7 @@
 }
 
 @book{zohdi2012electromagnetic,
-  author    = {Zohdi, Tarek I},
+  author    = {Zohdi, T. I.},
   doi       = {10.1007/978-3-642-28427-4},
   publisher = {Springer Science \& Business Media},
   title     = {Electromagnetic properties of multiphase dielectrics: a primer on modeling, theory and computation},
