### 2025-05-27 4209e9b
Fixing typesetting issues, other minor typo fixes
