+++ paper/markdown/paper.md
@@ -55,7 +55,7 @@ Project](https://next-gen.materialsproject.org/) database, ii)
 integration with the [Materials Project
 API](https://next-gen.materialsproject.org/api), iii) use of genetic
 machine-learning and iv) agnosticism to underlying microstructure, and
-ultimate engineering application, make it a tool with much broader
+v) ultimate engineering application, make it a tool with much broader
 applications than its predecessors.
 
 # Statement of need
@@ -65,8 +65,8 @@ enhanced material properties as compared to their individual
 constituents. As such, composite design is an active field, but the
 pursuit of new materials through experimentation is expensive. Today,
 computational tools for materials design are integral to reducing the
-cost and increasing the pace of innovation in the energy, electronics,
-aviation sectors, and beyond.
+cost and increasing the pace of innovation in sectors like energy, electronics,
+aviation and beyond.
 
 Several Python packages already exist for specific areas in composites
 modeling, such as
@@ -121,7 +121,7 @@ generalized effective material property $y^{*}$ of an $n$-phase
 composite. The generalized material properties for the $n$-phases are
 ordered from least to greatest where
 $y_{1} \leq y_{2} \leq \dotsb \leq y_{n}$ with corresponding volume
-fractions sum to unity $v_{1} + v_{2} + \dotsb + v_{n} = 1$:
+fractions that sum to unity $v_{1} + v_{2} + \dotsb + v_{n} = 1$, and are given by
 
 \begin{equation}\label{eqn:gen_ineq}
 y_{1} + \frac{A_{1}}{1 - \alpha_{1}A_{1}}
@@ -146,11 +146,11 @@ A_{n} = \sum\limits_{i=1}^{n-1} \frac{v_{i}}{\frac{1}{y_{i} - y_{n}} + \alpha_{n
 ![An example of a quasi-isotropic, quasi-homogeneous 3-phase
 composite. \label{fig:cartoon-3phase}](figures/3phase_composite.png){width=2in}
 
-\autoref{eqn:bulk_ineq) and 
+\autoref{eqn:bulk_ineq} and 
 \autoref{eqn:shear_ineq} summarize the results of the
 Hashin-Shtrikman derivations on the bounds on effective bulk modulus
-$\kappa^{*}$ and effective shear modulus $\mu^{*}$, where we
-simultaneously require
+$\kappa^{*}$ and effective shear modulus $\mu^{*}$, where it is
+simultaneously required that
 $\kappa_{1} \leq \kappa_{2} \leq \dotsb \leq \kappa_{n}$ and
 $\mu_{1} \leq \mu_{2} \leq \dotsb \leq \mu_{n}$:
 
@@ -206,8 +206,8 @@ expression for the effective property as
 y^{*} = \gamma y^{*,-} + (1-\gamma) y^{*,+}.
 \end{equation}
 
-Given experimental data we could fit $\gamma$ and extrapolate for a
-range of volume fractions, but, in the absence of experimental data, we
+Given experimental data, one could fit $\gamma$ and extrapolate for a
+range of volume fractions, but, in the absence of experimental data, the authors
 select $\gamma = 0.5$. For more information, the reader is referred to
 [@zohdi-course-reader].
 
@@ -268,7 +268,7 @@ subsequently given by
 
 and for phase 1 as
 \begin{equation}\label{eqn:C1load}
-\mbox{\boldmath $C$}_{1,\text{load}} = \frac{1}{v_{1}} \left( 1 - \sum\limits_{i=2}^{n} v_{i} \mbox{\boldmath $C$}_{i,\text{load}} \right),
[…]
