### 2025-06-15 #8209 @carlacupcake
Hi @prashjha, Here are some potential reviewers:

@giolajide <https://github.com/giolajide>
@vishsas <https://github.com/vishsas>
@Extraweich <https://github.com/Extraweich>
@mpvanderschelling <https://github.com/mpvanderschelling>
@MarDiehl <https://github.com/MarDiehl>
@ranzhengcode <https://github.com/ranzhengcode>
@JunOu <https://github.com/JunOu>
@nkhlbr <https://github.com/nkhlbr>
@PatrickJTaylor <https://github.com/PatrickJTaylor>
@amkrajewski <https://github.com/amkrajewski>
@moorepants <https://github.com/moorepants>
@aritraroy24 <https://github.com/aritraroy24>
@bdice <https://github.com/bdice>
@naik-aakash <https://github.com/naik-aakash>
@kingyin3613 <https://github.com/kingyin3613>
@vigneshS29 <https://github.com/vigneshS29>
@PythonFZ <https://github.com/PythonFZ>
@streeve <https://github.com/streeve>
@Yazhuo-Liu <https://github.com/Yazhuo-Liu>
@JosePizarro3 <https://github.com/JosePizarro3>
@dominikusbrian <https://github.com/dominikusbrian>
@Quantumstud <https://github.com/Quantumstud>

Thank you,
Carla

> On Jun 14, 2025, at 11:36 PM, Lars Pastewka ***@***.***> wrote:
>
>
> pastewka
>  left a comment
> (openjournals/joss-reviews#8209)
>  <https://github.com/openjournals/joss-reviews/issues/8209#issuecomment-2973535510>
> Happy to look at this @prashjha <https://github.com/prashjha> - but I have to warn that I am not the fastest these days (i.e. review time will likely be longer than a month).
>
