### 2025-07-16 9b07dc0
trying to fix broken example links in mkdocs
