### 2025-07-16 56d2e99
updated PyPI to 0.1.1 - fixed extraweich email from 7/12/2025
