### 2025-07-16 04ae513
fixed pandoc compilation errors
