### 2025-07-16 b3f3398
removed unnecessary paper.tex
