+++ paper/markdown/paper.tex
@@ -1,519 +1,0 @@
-\section{Summary}\label{summary}
-
-\texttt{hashin\_shtrikman\_mp} is a tool for composites designers who
-have desired composite properties in mind, but who do not yet have an
-underlying formulation. The library utilizes the tightest theoretical
-bounds on the effective properties of composite materials with
-unspecified microstructure -- the Hashin-Shtrikman bounds -- to identify
-candidate theoretical materials, find real materials that are close to
-the candidates, and determine the optimal volume fractions for each of
-the constituents in the resulting composite. Its i) leveraging of
-materials in the \href{https://next-gen.materialsproject.org/}{Materials
-Project} database, ii) integration with the
-\href{https://next-gen.materialsproject.org/api}{Materials Project API},
-iii) use of genetic machine-learning and iv) agnosticism to underlying
-microstructure, and v) ultimate engineering application, make it a tool
-with much broader applications than its predecessors.
-
-\section{Statement of need}\label{statement-of-need}
-
-Composites are ubiquitous in engineering due to their tunability and
-enhanced material properties as compared to their individual
-constituents. As such, composite design is an active field, but the
-pursuit of new materials through experimentation is expensive. Today,
-computational tools for materials design are integral to reducing the
-cost and increasing the pace of innovation in sectors like energy,
-electronics, aviation and beyond.
-
-Several Python packages already exist for specific areas in composites
-modeling, such as \href{https://github.com/Extraweich/homopy}{HomoPy},
-\href{https://github.com/rafaelcidade/compositeslib}{CompositesLib},
-\href{https://github.com/echaffey/Compysite}{Compysite},
-\href{https://github.com/azzeddinetiba/FeCLAP}{FeCLAP}, and
-\href{https://pypi.org/project/material-mechanics/}{material-mechanics},
-all of which perform stress analysis on laminates and/or
-fiber-reinforced composites using either classical laminate theory or
-the finite element method. Others exist which, like
-\texttt{hashin\_shtrikman\_mp}, utilize the Hashin-Shtrikman bounds on
-effective composite properties, such as
-\href{https://geodynamics.github.io/burnman/}{BurnMan} for thermal
-analysis of composite rocks/assemblages,
-\href{https://rockphypy.readthedocs.io/en/latest/getting_started/08_Shaly_sand_modelling.html}{rockphypy}
-for mechanical modeling of sand-shale systems, {[}@ZARE2017176{]}'s
-modeling of clay nanocomposites,
-\href{https://github.com/JulianKarlBauer/mechmean/tree/main?tab=readme-ov-file}{mechmean},
-and {[}@ZERHOUNI2019344{]}'s modeling of 3D printed microstructures. All
-of these tools, however, are highly specific to composite
-microstructure, macro-geometry, and composition. More notably, they
-focus on analysis of already well-defined composites, rather than
-discovery of new materials.
-
[…]
