### 2025-07-16 f235022
fix figure sizes
