### 2025-07-16 e4fe665
small update to eqn 28
