### 2025-07-16 #8412 @carlacupcake
Hi @Extraweich ,

Thank you so much for reviewing. I believe I have addressed the comments above and fixed the .yaml FileNotFoundError you mentioned in a different email on July 12th. Please let me know if you find other issues!
Please try `pip install hashin-shtrikman-mp==0.1.1`

Best,
Carla
