### 2025-08-02 d87f1ce
Update CODE_OF_CONDUCT.md
