### 2025-08-04 #8412 @Extraweich
Hi @carlacupcake,

let me try to rephrase my comments:

- Given two materials, I can go ahead and calculate a property, e.g., the bulk modulus, of the composite made of those two materials for any volume fraction $c\in[0,1]$, using the upper/lower Hashin-Shtrikman bounds. The only free parameter would be the volume fraction $c$. Neverthless, in your example you feed those two materials into your optimiser (genetic algorithm). Here I don't quite understand what the genetic algorithm is doing, because I could calculate the effective properties without using any optimiser. I guess, I am not seeing something here.
- The other issue that I have is that I would want to know the effective composite properties based on the Hashin-Shtrikman bounds for the list of best fits. Is there an option that the `MatchFinder` tells me what the effective bulk modulus would be for every item in the list, so I can compare them with my desired value?

Thanks for adding the CODE_OF_CONDUCT.md and thanks for highlighting your ~/tests directory, which I had overseen.

Besides the two open questions above, for which I'm sure we'll come to an agreement, I see great value in the project and would recommend it to be published in JOSS.
