### 2025-08-04 #8412 @carlacupcake
Hi @Extraweich,

I understand your confusion now and concede the example could be simplified, but perhaps you will accept this explanation.

- Your understanding is completely correct for the example -- volume fraction is the only free parameter in the optimization, which begs the question, why not just do a brute force grid search at that point? We could, but the true answer is I didn't want to right more code and the genetic algorithm will get you to the same result. I can write the grid search into the example if you think that would add clarity.
- I have added the effective properties to the summary table, which you can see in `example_n=3.ipynb`. This was a great suggestion, thank you.

I'm glad to hear you see value in the project!
