### 2025-08-07 #8412 @carlacupcake
Hi @Extraweich,

In the output of cell [11] in `example_n=3.ipynb` you should be able to see that the final 6 columns are now the 5 effective properties of the composite and the cost. Please let me know if you don't see those columns in the latest commit, or if you think I should make them more prominent.

Best,
Carla
