### 2025-08-07 #8412 @carlacupcake
By `Do you think a similar list could be visualized on the closest fits of the real material?` I would say the answer is yes, because for that one combination, you could run it through the optimizer with only volume fraction as a free variable, then print out the table of best designs, where the top row would be the "best" volume fraction combo.

Also, the pip version should now be up to date with the git version.
