### 2025-08-26 #8412 @pastewka
Hi @carlacupcake, sorry for the long time this has taken. Here are my comments on paper and code. Happy to discuss:

# Major in paper

1) The Harshin-Shtrikman bounds are valid for homogeneization of a specific elliptic differential equations and thereby valid for elasticity, heat-transport and other classes of properties that can be described by this equation. It may be useful to state what type homogeneization problem the bounds apply to before given the bounds themselves, i.e. give the governing equations. This may make some of what follows easier, see following comments.

2) The section "Quantifying distributed loads" is confusing; I would much prefer to have equations use symbols rather than descriptive text. Eq. (14) seem to use one symbol (C) but in Eq. (11) this is text although I believe those must be the same quantity.

3) Please also indicate what is a tensor through notation. I understand that the constitutive law may not be of a tensorial character, which may make consistent notation more complicated.

4) Eq. (11) is (I believe, but unclear) the constitutive response of the homogeneization problem. If the problem was introduced properly (see my comment 1 above) then this could section could refer to the problem statement which I believe would make the whole paper much clearer.

5) I don't understand the discussion around Eq. (10). First, gamma should be introduced. Second, I believe what is said is that the actual property is between the bound and gamma is an effective mixing parameter. The authors choose 0.5, but it is unclear to me at this point in this paper why this choice is actually necessary. From what is said before the objective is to get bounds, not the actual property (which of course depends on details of the microstructure).

# Minor in paper

6) DOI for Zohdi and Wriggers is wrong

7) Line 125: Reference to figure 2 is missing

# Code

8) Type annotations: Indicate optional results, e.g. line 20 of population.py turn `values: np.ndarray = None` into `values: Optional[np.ndarray] = None` (since you seem to target Python 3.9 as a minimum version). There are more instances where this should be done. If you are fine to limit to Python 3.10 `values: np.ndarray | None = None` would probably be better.

9) What is the data in io/outputs folder? It does not appear to be used anywhere and I would suggest to remove it if that is actually the case.

# Curiosity (you may want to comment on this in the paper)

10) Why do you need a genetic algorithm and cannot just use a gradient-based optimizer? (Are there too many local minima and you are looking for the global minimum?)
