### 2025-09-20 f55469d
fixed lars comments about type annotations and io/output
