+++ paper/markdown/paper.bib
@@ -79,7 +79,7 @@ year    = {2019},
 
 @book{zohdi-course-reader,
   author    = {Zohdi, T. I. and Wriggers, P.},
-  doi       = {10.1007/11672913},
+  doi       = {10.1007/978-3-540-32360-0},
   publisher = {Springer-Verlag},
   title     = {Introduction to computational micromechanics},
   year      = {2008},
