### 2025-09-21 66f0fea
"Minor in paper" and "Code" from Lars addressed
