### 2025-09-21 c342bc0
addressed "Major in paper" and "Curiosity" from Lars
