+++ paper/markdown/paper.md
@@ -105,7 +105,24 @@ constituents.
 
 # Underlying theory
 
-## Estimate effective composite properties with the Hashin-Shtrikman bounds
+## Governing equations and applicability of the Hashin-Shtrikman bounds
+The Hashin–Shtrikman variational bounds apply to homogenization problems 
+governed by linear, self-adjoint, elliptic partial differential equations 
+of the form
+
+\begin{equation}\label{eqn:ellipstic}
+-\nabla\cdot(\boldsymbol{A}:(\boldsymbol{x})\nabla u(\boldsymbol{x})) = f(\boldsymbol{x}),
+\end{equation}
+
+where $\boldsymbol{A}(\boldsymbol{x})$ is a symmetric positive-definite local property tensor. 
+This general formulation covers a wide class of physical phenomena including 
+steady-state heat conduction ($\boldsymbol{A} = \boldsymbol{k}(x)$, thermal conductivity), electrostatics 
+($\boldsymbol{A} = \boldsymbol{\sigma}_{e}(\boldsymbol{x})$, electrical conductivity), and linear elasticity ($\boldsymbol{A} 
+= \mathbb{E}(\boldsymbol{x})$, stiffness). Under these assumptions, the effective "proportionality" 
+tensor $\boldsymbol{A}^{*}$ of the composite is uniquely defined, and the Hashin–Shtrikman bounds 
+give the tightest possible estimates on given only the phase properties and volume fractions.
+
+## Estimate effective composite properties with the HS bounds
 
 When designing composites, simple volume-weighted linear combinations of
 constituent material properties do not yield accurate approximations of
@@ -116,10 +133,10 @@ Hill-Reuss-Voight-Weiner bounds, where the lower bound is the harmonic
 mean of the constituent material properties and the upper bound is the
 arithmetic mean [@commentaryHS]. For quasi-isotropic and
 quasi-homogeneous multi-phase composites with arbitrary phase geometry
-(a more general case), a better option is to use the Hashin-Shtrikman
+(a more general case), a better option is to use the HS
 bounds, which provide even tighter ranges on the resulting effective
 properties [@hashin1962variational]. \autoref{eqn:gen_ineq}
-summarizes the Hashin-Shtrikman bounds on a
+summarizes the HS bounds on a
 generalized effective material property $y^{*}$ of an $n$-phase
 composite. The generalized material properties for the $n$-phases are
 ordered from least to greatest where
@@ -151,7 +168,7 @@ composite. \label{fig:cartoon-3phase}](figures/3phase_composite.png){width=2in f
 
 \autoref{eqn:bulk_ineq} and 
 \autoref{eqn:shear_ineq} summarize the results of the
-Hashin-Shtrikman derivations on the bounds on effective bulk modulus
+HS derivations on the bounds on effective bulk modulus
 $\kappa^{*}$ and effective shear modulus $\mu^{*}$, where it is
 simultaneously required that
 $\kappa_{1} \leq \kappa_{2} \leq \dotsb \leq \kappa_{n}$ and
@@ -225,47 +242,47 @@ failure. By introducing the concept of "concentration tensors", we can
 quantify a constituent's contribution to load response and then use
 constitutive laws to determine how the composite will respond to loads.
 
-For the general case of a constitutive law of the form
-\begin{equation}\label{eqn:gen-const-law}
[…]
