+++ paper/markdown/paper.md
@@ -106,11 +106,11 @@ constituents.
 # Underlying theory
 
 ## Governing equations and applicability of the Hashin-Shtrikman bounds
-The Hashin–Shtrikman variational bounds apply to homogenization problems 
+The Hashin–Shtrikman (HS) variational bounds apply to homogenization problems 
 governed by linear, self-adjoint, elliptic partial differential equations 
 of the form
 
-\begin{equation}\label{eqn:ellipstic}
+\begin{equation}\label{eqn:elliptic}
 -\nabla\cdot(\boldsymbol{A}:(\boldsymbol{x})\nabla u(\boldsymbol{x})) = f(\boldsymbol{x}),
 \end{equation}
 
@@ -120,7 +120,7 @@ steady-state heat conduction ($\boldsymbol{A} = \boldsymbol{k}(x)$, thermal cond
 ($\boldsymbol{A} = \boldsymbol{\sigma}_{e}(\boldsymbol{x})$, electrical conductivity), and linear elasticity ($\boldsymbol{A} 
 = \mathbb{E}(\boldsymbol{x})$, stiffness). Under these assumptions, the effective "proportionality" 
 tensor $\boldsymbol{A}^{*}$ of the composite is uniquely defined, and the Hashin–Shtrikman bounds 
-give the tightest possible estimates on given only the phase properties and volume fractions.
+give the tightest possible estimates on $\boldsymbol{A}^{*}$ given only the phase properties and volume fractions.
 
 ## Estimate effective composite properties with the HS bounds
 
@@ -242,15 +242,17 @@ failure. By introducing the concept of "concentration tensors", we can
 quantify a constituent's contribution to load response and then use
 constitutive laws to determine how the composite will respond to loads.
 
-For the general case of a constitutive law of the form \autoref{eqn:elliptic}, which can also be written more conveniently as
+The general constitutive law given in \autoref{eqn:elliptic} can also be written more conveniently as
 \begin{equation}\label{eqn:easy-elliptic}
 \boldsymbol{\xi}(\boldsymbol{x}) = \boldsymbol{A}(\boldsymbol{x}):\nabla u(\boldsymbol{x}),
 \end{equation}
 
-over the domain $\Omega$, we note that, by the definition of volume
+where $\nabla u(\boldsymbol{x})$ is a tensor-valued load (a field gradient like strain, electric field, or temperature gradient), $\boldsymbol{\xi}(\boldsymbol{x})$ is a tensor-valued response (a flux like stress, current, or heat flux), and $\boldsymbol{A}(\boldsymbol{x})$ is a proportionality tensor (stiffness, conductivity).
+
+Over the domain $\Omega$, we note that, by the definition of volume
 fraction, the tensor-valued load will abide by:
 \begin{equation}\label{eqn:by-vol-frac-load}
[…]
