### 2025-09-21 cf247f4
fixed double superscript compilation error
