### 2025-09-21 efb5c72
actually fixing double superscript
