### 2025-09-21 61e98b2
proportionality -> local property
