### 2025-09-21 #8412 @carlacupcake
Hello @prashjha and @pastewka,

Apologies for the late reply -- this past month has been hectic.

Lars, I appreciate your detailed read of both the paper and the code base and have taken many of your suggestions. I will address each of them below and note where in the paper/code I changed things.

# Major in paper
1. I have added a section entitled "Governing equations and the applicability of the Hashin-Shtrikman bounds" which notes the HS apply to elliptic DEs.
2. Given the symbols used in the general governing equation from "Governing equations and the applicability of the Hashin-Shtrikman bounds", I changed all the descriptive text to symbols.
3. I tried to use bolding and double-dots to better indicate tensors throughout.
4. What you noted as Eq. 11 is now Eq. 12. It is (hopefully) a clearer statement of the general linear constitutive law that comes from the elliptic DE in "Governing equations and the applicability of the Hashin-Shtrikman bounds".
5. This is where I did not change anything. Maybe the following explanation will help. The actual property will be in between the bounds and, yes, gamma is a mixing parameter. I thought this followed from Eq. 11 without explicit statement, but I can explicitly state it if you suggest. We eventually want to get from bounds to an estimate for the property for the purposes of the optimization. Do you suggest I move the discussion about gamma to the cost function section?

# Minor in paper
6. Should be fixed.
7. Should be fixed. Now line 138.

# Code
8. I do indeed target Python 3.9 as the minimum version. Accordingly, I changed all type annotations with `= None` to use `Optional`.
9. Those data are the dictionaries of materials and their properties pulled from the MP database upon an API request. I changed the code to remove the files after usage by the match finder.

# Curiosity
10. I added this sentence to the final paragraph of the cost function section "Genetic algorithms are great for exploring promising parts of the design space, and since the end goal is to choose composite constituents from a set of discrete choices (materials in the MP database), no further refinement with a gradient-based optimization method is needed." Please let me know if this does not address your question.
