### 2025-09-30 3faf97b
trying to fix the spacing issue with "good" ln 170
