### 2025-10-06 #8412 @carlacupcake
Hi @prashjha,

What are the next steps in the publication process?

Thank you,
Carla

On Wed, Oct 1, 2025 at 9:17 AM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#8412)
> <https://github.com/openjournals/joss-reviews/issues/8412#issuecomment-3357148660>
>
> 👋 @openjournals/bcm-eics
> <https://github.com/orgs/openjournals/teams/bcm-eics>, this paper is
> ready to be accepted and published.
>
> Check final proof 👉📄 Download article
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.08412/joss.08412/10.21105.joss.08412.pdf>
>
> If the paper PDF and the deposit XML files look good in
> openjournals/joss-papers#7239
> <https://github.com/openjournals/joss-papers/pull/7239>, then you can now
> move forward with accepting the submission by compiling again with the
> command @editorialbot accept
>
