### 2025-10-07 #8412 @carlacupcake
Hi @Kevin-Mattheus-Moerman

I have made the USA -> United States of America change.

For the additional point about the length of the paper, would you say it is a hard requirement to shorten the paper? My advisor and co-author, Tarek Zohdi, is intent on keeping both the Underlying Theory and Cost Function sections.

That said, if I were to move that content to a supplemental markdown file, is there a JOSS md template and GitHub associated GitHub action?

Thank you,
Carla
